Don't use unsafe FFI call for ptrace!
As per GHC user guide, unsafe calls must not be used for blocking
IO functions -- they block the runtime!

Note though that ptrace() invocations must be made from the OS *thread*
that is the tracer, see
    https://stackoverflow.com/questions/18737866/can-a-thread-trace-a-process/18759305#18759305
and what the module haddocks already say.
However, before this commit, using `forkOn` in combination with the `unsafe`
`c_ptrace` was enough to enforce this. With this commit changing it to `safe`,
this is no longer the case, and the `runInBoundThread` or `forkOS` that were
already mentioned in the module haddocks is now the only known way to enforce
it; `forkOn` no longer works.